+++ src/test/kotlin/no/nav/infotrygd/beregningsgrunnlag/dto/ConvertersKtTest.kt
@@ -93,7 +93,7 @@ class ConvertersKtTest {
         val stoenadstype = Stoenadstype.FOEDSEL // ytelse = foreldrepenger
         val opprinneligIdentdato = LocalDate.now()
         val dekningsgrad = 75
-        val gradering = 50
+        val gradering = 50.toBigDecimal()
         val foedselsdatoBarn = LocalDate.now().minusYears(1)
 
         val pf = TestData.PeriodeFactory()
